+++ spec/units/gandi_v5/domain_spec.rb
@@ -128,13 +128,13 @@ describe GandiV5::Domain do
 
   describe '.create' do
     let(:url) { 'https://api.gandi.net/v5/domain/domains' }
+    let(:returns) { double described_class }
+    let(:response) { double RestClient::Response, headers: { location: 'https://api.gandi.net/v5/domains/example.com' } }
 
     describe 'Sets dry-run header' do
       let(:body) { '{"owner":{},"fqdn":"example.com"}' }
 
       it 'False by default' do
-        returns = double described_class
-        response = double RestClient::Response, headers: { location: 'https://api.gandi.net/v5/domains/example.com' }
         expect(GandiV5).to receive(:post).with(url, body, 'Dry-Run': 0).and_return([response, nil])
         expect(described_class).to receive(:fetch).with('example.com').and_return(returns)
         described_class.create 'example.com', owner: {}
@@ -146,8 +146,6 @@ describe GandiV5::Domain do
       end
 
       it 'False' do
-        returns = double described_class
-        response = double RestClient::Response, headers: { location: 'https://api.gandi.net/v5/domains/example.com' }
         expect(GandiV5).to receive(:post).with(url, body, 'Dry-Run': 0).and_return([response, nil])
         expect(described_class).to receive(:fetch).with('example.com').and_return(returns)
         described_class.create 'example.com', owner: {}, dry_run: false
@@ -171,12 +169,29 @@ describe GandiV5::Domain do
       end
     end
 
+    describe 'Sets sharing_id' do
+      it 'Absent by default' do
+        expect(GandiV5).to receive(:post).with(url, any_args).and_return([response, nil])
+        expect(described_class).to receive(:fetch).with('example.com').and_return(returns)
+        described_class.create('example.com', owner: {})
+      end
+
+      it 'Paying as another organization' do
+        expect(GandiV5).to receive(:post).with("#{url}?sharing_id=organization_id", any_args).and_return([response, nil])
+        expect(described_class).to receive(:fetch).with('example.com').and_return(returns)
+        described_class.create('example.com', sharing_id: 'organization_id', owner: {})
+      end
+
+      it 'Buy as a reseller' do
+        expect(GandiV5).to receive(:post).with("#{url}?sharing_id=reseller_id", any_args).and_return([response, nil])
+        expect(described_class).to receive(:fetch).with('example.com').and_return(returns)
+        described_class.create('example.com', sharing_id: 'reseller_id', owner: {})
+      end
+    end
+
     it 'Success' do
-      returns = double described_class
-      response = double RestClient::Response, headers: { location: 'https://api.gandi.net/v5/domains/example.com' }
       body = '{"owner":{},"fqdn":"example.com"}'
-      expect(GandiV5).to receive(:post).with(url, body, 'Dry-Run': 0)
-                                       .and_return([response, nil])
+      expect(GandiV5).to receive(:post).with(url, body, 'Dry-Run': 0).and_return([response, nil])
       expect(described_class).to receive(:fetch).with('example.com').and_return(returns)
       expect(described_class.create('example.com', owner: {})).to be returns
     end
@@ -186,18 +201,14 @@ describe GandiV5::Domain do
     end
 
     it 'Given contact as hash' do
-      returns = double described_class
       body = '{"owner":{"email":"owner@example.com"},"fqdn":"example.com"}'
-      response = double RestClient::Response, headers: { location: 'https://api.gandi.net/v5/domains/example.com' }
       expect(GandiV5).to receive(:post).with(url, body, 'Dry-Run': 0).and_return([response, nil])
       expect(described_class).to receive(:fetch).with('example.com').and_return(returns)
       described_class.create 'example.com', owner: { email: 'owner@example.com' }
     end
 
     it 'Given contact as GandiV5::Domain::Contact' do
-      returns = double described_class
       body = '{"owner":{"email":"owner@example.com"},"fqdn":"example.com"}'
-      response = double RestClient::Response, headers: { location: 'https://api.gandi.net/v5/domains/example.com' }
       expect(GandiV5).to receive(:post).with(url, body, 'Dry-Run': 0).and_return([response, nil])
       expect(described_class).to receive(:fetch).with('example.com').and_return(returns)
       owner = double GandiV5::Domain::Contact, to_gandi: { 'email' => 'owner@example.com' }